% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxstripchart.R
\name{boxstripchart}
\alias{boxstripchart}
\title{Plot boxplot together with stripchart dotplot.}
\usage{
boxstripchart(y, x, ...)
}
\arguments{
\item{y}{values}

\item{x}{groups}

\item{...}{will be parsed to boxplot and stripchart}
}
\value{
nothing
}
\description{
Plot boxplot together with stripchart dotplot.
}
\author{
Wenjian Yang
}
