% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxstripchart.R
\name{boxviolin}
\alias{boxviolin}
\title{Plot boxplot together with violin dots}
\usage{
boxviolin(
  y,
  x,
  srt = 0,
  xtext.adj = NULL,
  box.more.args = list(),
  dots.more.args = list(),
  cex.axis = 0.8,
  ...
)
}
\arguments{
\item{y}{variable to plot}

\item{x}{group variable}

\item{srt}{angle of x-axis annotation}

\item{xtext.adj}{placement of x-axis nnotation}

\item{box.more.args}{list of args for boxplot}

\item{dots.more.args}{list of args for dots, col/pch are by group}

\item{cex.axis}{size of axis labels}

\item{...}{additional argments to parse down to boxplot and viopoints}
}
\value{
none
}
\description{
Plot boxplot together with violin dots
}
\author{
Wenjian Yang
}
