% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_func.R
\name{format_pvalue_table}
\alias{format_pvalue_table}
\title{Format pvalues in a table}
\usage{
format_pvalue_table(x, digits = 3)
}
\arguments{
\item{x}{table}

\item{digits}{digits to keep for p-values}
}
\value{
a table with proper digits for p-values
}
\description{
Look for columns that representing p-values, range 0 to 1;
if p-value is less than 1, signif(pvalue, digits);
else, round(pvalue, digits).
}
\author{
Wenjian Yang
}
