% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_conf_int}
\alias{get_conf_int}
\title{Compute Confidence intervals from estimates and p-values}
\usage{
get_conf_int(ests, pvalues, ci = 0.95, exponential = FALSE)
}
\arguments{
\item{ests}{a vector of coefficients of estimate}

\item{pvalues}{a vector of p-values}

\item{ci}{confidence interval, default 0.95}

\item{exponential}{whether take the exponential transformation for the coef, default FALSE}
}
\value{
a matrix with est, 95\%CI and pvalue
}
\description{
Compute Confidence intervals from estimates and p-values based on normal distribution
}
\author{
Wenjian Yang
}
