% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{get_file_header}
\alias{get_file_header}
\title{retrieve header lines from a text file}
\usage{
get_file_header(filename, limit = 5000, header.char = "#")
}
\arguments{
\item{filename}{path to the file}

\item{limit}{number of lines to check for header, default 5000}

\item{header.char}{default '#'}
}
\value{
a vector and each line is an alement
}
\description{
read header lines from a text file
assuming header starts with '#'
}
\author{
Wenjian Yang
}
