% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_odds_ratio}
\alias{get_odds_ratio}
\title{extract odds ratio/hazards ratio from glm or coxph model}
\usage{
get_odds_ratio(m)
}
\arguments{
\item{m}{model}
}
\value{
a matrix with est, 95\%CI and pvalue
}
\description{
Compute Odds Ratio/Hazards ratio based on estimate and standard error
}
\author{
Wenjian Yang
}
