% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_func.R
\name{stringify.table}
\alias{stringify.table}
\title{Concise print a table}
\usage{
stringify.table(tt, row.sep = ",", col.sep = "|")
}
\arguments{
\item{tt}{table}

\item{row.sep}{row_separator, default ","}

\item{col.sep}{column_separator, default "|"}
}
\value{
a string
}
\description{
create a string representing a table,
with default row_separator "," and default column_separator "|"
}
\author{
Wenjian Yang
}
