% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.col.R
\name{trans.col}
\alias{trans.col}
\title{Transparent color}
\usage{
trans.col(col, alpha = 1)
}
\arguments{
\item{col}{color in string}

\item{alpha}{transparency 0-1}
}
\value{
rgb for transparent color
}
\description{
copied from the internet
}
\author{
Yang
}
