% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_func.R
\name{format_pvalue_table}
\alias{format_pvalue_table}
\title{Format pvalues in a table}
\usage{
format_pvalue_table(x, digits = 3)
}
\arguments{
\item{x}{table}

\item{digits}{digits to keep for p-values}
}
\value{
a table with proper digits for p-values
}
\description{
Look for columns that representing p-values, range 0 to 1;
if p-value is less than 1, signif(pvalue, digits);
else, round(pvalue, digits).
}
\examples{
tt <- data.frame(v1=c('A', 'B', 'C'), v2=c(0.000003, 0.002131, 0.442132), v3=c(-2.0, 0.5, 1.123))
format_pvalue_table(tt, 2)
}
\author{
Wenjian Yang
}
