% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.normal.transform.R
\name{rank.normal.transform}
\alias{rank.normal.transform}
\title{rank.normal.transform}
\usage{
rank.normal.transform(x)
}
\arguments{
\item{x}{numeric vector to be rank-normal-transformed}
}
\value{
transformed values following normal distribution
}
\description{
Inverse rank normal transformation for quantative trait
}
\details{
Inverse rank normal transofrmation
}
\author{
Wenjian Yang
}
