% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_func.R
\name{stringify.table}
\alias{stringify.table}
\title{Concise print a table}
\usage{
stringify.table(tt, row.sep = ",", col.sep = "|")
}
\arguments{
\item{tt}{table}

\item{row.sep}{row_separator, default ","}

\item{col.sep}{column_separator, default "|"}
}
\value{
a string
}
\description{
create a string representing a table,
with default row_separator "," and default column_separator "|"
}
\examples{
tt <- data.frame(v1=c('A', 'B', 'C'), v2=c(0.3, 0.002131, 0.442132), v3=c(-2.0, 0.5, 1.123))
stringify.table(tt)

}
\author{
Wenjian Yang
}
